fixed an error on the db, merged the two db classes into one and integrated that with the other entity classes
