+++ src/main/resources/sql/setup.sql
@@ -71,5 +71,5 @@ CREATE TABLE IF NOT EXISTS Invoice (
 id INTEGER PRIMARY KEY AUTOINCREMENT,
 order_id INTEGER NOT NULL,
 total_payable_amount REAL NOT NULL,
-FOREIGN KEY (order_id) REFERENCES OrdersStatus(id) ON DELETE CASCADE
+FOREIGN KEY (order_id) REFERENCES Orders(id) ON DELETE CASCADE
 );
